Allow reading reading the set of meshes (together with their points/normals - object data) from the scene.
TODO: this gives the user the ability to put the scene into an inconsistent state if he e.g insert points/normals which are not part of any triangle or if he insert a new objData_t with a material that is not in the materials list etc. To fix this while preserving the performance it would be better to create a visitor for the triangles in the scene.